Junit fixed a failing test case

+++ PHASE3/junit_demo/src/main/java/junit_demo/Calculator.java
@@ -4,7 +4,7 @@ public class Calculator {
 	
 	long add(int a, int b) {
 		
-		return a+b;
+		return (long) (a + b);
 		
 		//throw new RuntimeException("Method not implemented");
 	}
